// generated/src/app/api-models/group-contact-person.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GroupContactPerson {
  seqGroupContact: number;
  seqGroupId: number;
  contactName: string;
  contactTitle: string;
  phoneNumber: string;
  extension: string;
  faxNumber: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  primaryDistributionMethod: string;
  emailId: string;
}